Separate JPA and Mongo repos
