+++ sensor-api/src/main/java/com/example/sensorapi/repository/MongoSensorReadingStore.java
@@ -1,6 +1,7 @@
 package com.example.sensorapi.repository;
 
 import com.example.sensorapi.model.SensorReading;
+import com.example.sensorapi.repository.mongo.SensorReadingMongoRepository;
 
 import java.util.List;
