+++ sensor-api/src/main/java/com/example/sensorapi/repository/jpa/SensorReadingRepository.java
@@ -1,0 +1,7 @@
+package com.example.sensorapi.repository.jpa;
+
+import com.example.sensorapi.model.SensorReading;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SensorReadingRepository extends JpaRepository<SensorReading, Long> {
+}
