+++ sensor-api/src/main/java/com/example/sensorapi/repository/mongo/SensorReadingMongoRepository.java
@@ -1,4 +1,4 @@
-package com.example.sensorapi.repository;
+package com.example.sensorapi.repository.mongo;
 
 import com.example.sensorapi.model.SensorReading;
 import org.springframework.data.mongodb.repository.MongoRepository;
